Cloud-config support
Provide a means to manipulate user-data for cloud config, including
support for MIME multipart, cloud config archive,
and raw shell scripts.

Specifically, allow a new section to be prepended to whatever exists,
preserving the original functionality but potentially altering the
method in which it is stored in user-data.

Note that DigitalOcean does not support base64-encoded gzipped user-data
so there is no provision to perform these operations.