secure introduction
Allow new droplets to be provided with a personalised
request-wrapped single-use SecretID issued by Vault,
which is only usable by their IP address for a few minutes.

This increases the ability to track which droplets are which, and
avoids exposing an exploitable token to third parties which can
access droplet metadata.