Allow duration of SecretID and the request wrapper to be different
While the request-wrapped SecretID is one-shot and intended to be
used during cloud-init, exchanging it for the actual SecretID,
the resultant SecretID's validity may need to be greater than this.

Some use-cases will have the droplet quickly exercise the SecretID to
retrieve general-purpose secrets, but other use-cases will have the
droplet continue to use the AppRole/SecretID for the entirety of the
droplet's lifetime.