clean up orphans on scale-out
If a droplet is created by the autoscaler, but it fails to
register with nomad as a client, it should be considered
an orphan and deleted.

Add a configuration setting to control this behaviour. Only
droplets with the "name" tag which were not recently created
will be considered for deletion.